debate_app/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
google-genai
google-api-core
python-dotenv
